+++ extensions/wikia/FandomCreatorEmail/FandomCreatorEmail.setup.php
@@ -1,0 +1,5 @@
+<?php
+
+$wgAutoloadClasses['FandomCreatorEmail\MagicWordWrapper'] = __DIR__ . '/MagicWordWrapper.php';
+$wgAutoloadClasses['FandomCreatorEmail\FandomCreatorEmailController'] = __DIR__ . '/FandomCreatorEmailController.php';
+$wgAutoloadClasses['FandomCreatorEmail\Controller\ContentUpdatedController'] = __DIR__ . '/Controller/ContentUpdatedController.php';
