+++ extensions/wikia/FandomCreatorEmail/MagicWordWrapper.php
@@ -1,7 +1,6 @@
 <?php
 
-namespace Email\Controller\FandomCreator;
-
+namespace FandomCreatorEmail;
 
 use Wikia\Util\GlobalStateWrapper;
